Доработана авторизация, добалена клавиатура ввода пина

+++ src/main/groovy/friend/bot/entity/Message.groovy
@@ -9,7 +9,7 @@ public class Message {
 
 	@Id
 	@GeneratedValue
-	Integer id
+	Long id
 	
     String text
