+++ src/main/groovy/friend/bot/service/UserService.groovy
@@ -13,13 +13,13 @@ import org.slf4j.LoggerFactory;
 class UserService {
 	private static final Logger log = LoggerFactory.getLogger(UserService.class);
 	
-	private String getUserEmail(Integer userId) {
+	private String getUserEmail(Long userId) {
 		
 	}
 	
 	
 	
-	public FriendUser Auth(Integer userId) {
+	public FriendUser Auth(Long userId) {
 		log.debug("Проверяем ИД пользователя: {}",userId);
 		return new FriendUser(userId);
 	}
